initial systemd-repart configuration for /boot/efi
The mkosi default systemd-repart configuration mounts the ESP on /boot, which
is not what Debian expects. This change introduces customized systemd-repart
configuration that mounts the ESP on /boot/efi as expected.

Note that this is still not exactly right; all the contents of /boot are
currently copied to the ESP, which is not the desired configuration and likely
breaks BIOS boot with GRUB.